src/services/rag_service/requirements.txt
llama-index==0.12.49
llama-index-embeddings-huggingface==0.5.5
llama-index-vector-stores-milvus==0.8.5
transformers==4.53.2
fastapi==0.116.1
uvicorn==0.35.0
pyyaml